Fix EvaluatePath resulting in huge types
